sync_manager: lease_threads handle watchdog files
I don't believe the complexity of using a separate thread to
update watchdog files for all leases is worth it.  Having each
lease thread update its watchdog file is far simpler.  The original
intention of a watchdog thread was to avoid the possibility that
lease renewal could be blocked by local fs access, but that's both
unlikely and not a safety risk.