sync_manager: reacquire resources
When leases are released for a pid that's still running, keep
the resource (along with its version) around.  If the resource
is later reacquired, check that the ondisk lease version matches
the version that was previously saved.  If they don't match, the
(re)acquire fails.

Set reacquire to 1 in the token arg passed to acquire to enable
the version verification.  (No command line option for setting
reacquire=1 yet.)