sync_manager: same pid may reacquire resource
If a pid releases a token on a resource, due to being paused
for example, that same pid is allowed to reacquire a token
on the same resource (off the deleted_resources list) when
it's resumed.  A different pid is not allowed to acquire a
token on a resource from the deleted_resources list since a
running pid will expect no one else to have owned the resource
between the time it was paused and resumed.

That is the local check.  Once passed, the distributed test
is enforced by checking that the disk lease version has not
changed since the lease was released.

This also means that when a pid exits, we must purge any
resources from the deleted_resources list that it may have
released while still running.