sanlock: recovery fixes
The relationships between delta_lease timeouts, paxos_lease timeouts
and watchdog timeouts are complicated and critical for safety.

We need to keep track of a host_id generation, so we can distinguish
between a lease acquired by a host before it failed once it returns.
This involves a disk format change.