sanlock: introduce lockspaces
A major reworking to remove the assumption that a host has a single
host_id in a single host_id lease area, all resources are owned by
that host_id, and if that host_id is not renewed, everything is
shut down.

That does not work with the rhev "storage domain" model where a host
can use multiple independent storage domains, and needs to maintain a
separate host_id within each.  Resources are associated with one of
the multiple storage domains, and if a storage domain is lost, only
resources and pids using it need to be shut down.

This uses a "lockspace" to support the "storage domain" concept.
Lockspaces are added and removed administratively, using the
"sanlock add_lockspace|rem_lockspace" commands.  Each resource
definition now begins with the name of the lockspace is lives
under.

Two major command line changes are how lockspaces and resources
are specified, -r RESOURCE and -s LOCKSPACE.  From the -h output:

LOCKSPACE = <lockspace_name>:<host_id>:<path>:<offset>
  <lockspace_name>	name of lockspace
  <host_id>		local host identifier in lockspace
  <path>		disk path where host_id leases are written
  <offset>		offset on disk, in bytes

RESOURCE = <lockspace_name>:<resource_name>:<path>:<offset>[:<path>:<offset>...]
  <lockspace_name>	name of lockspace
  <resource_name>	name of resource being leased
  <path>		disk path where resource leases are written
  <offset>[s|KB|MB]	offset on disk, default unit bytes
                        [s = sectors, KB = 1024 bytes, MB = 1024 KB]
  [:<path>:<offset>...] other disks in a multi-disk lease