sanlock: add lockspace checks
When comparing lockspaces, do a full compare of name, and
host_id path/offset/id when possible.  This prevent adding
two lockspaces with the same name but different host_id info.

(Of course two lockspaces with the same never should never
be created in the first place.  We can't fully protect
against that, so it's a user error, but we can reject it
when we see it.)