sanlock: renewal changes
Remove the pthread cond var and cond_timedwait and
replace with a much simpler sleep.  This seems to
eliminates the erratic renewal behavior, in addition
to making the code simpler.

In the renewal thread, only use the leader timestamp
when the renewal was successful.

Add/change debugging messages related to renewls.

Use local variables in renewal thread to guard against
corrupting the disk if a bug causes the sp struct be
corrupted.