sanlock: add flag RESTRICT_SIGKILL
which causes sanlock to only use SIGTERM to kill pid,
not to escalate to SIGKILL.  A process may be holding
a lease based on state outside itself; the fact that
the pid exits does not imply it's safe to release
the lease.  The pid would choose to ignore the SIGTERM
if it's not safe to release the lease, and using
SIGKILL makes it impossible for the pid to make that
choice.