sanlock: change to paxos algorithm
I don't think the way were were choosing the
mbal value was correct or consistent with the
official published description.

We now read the entire lease area at the start of
the paxos run and look through all dblocks to
find the maximum mbal, and base the new mbal on
that, rather than just basing it on our own previous
mbal value.

The paper isn't very clear about this, and the use
of lver in our implementation has a similar role to
mbal, and also makes things a bit different from the
published algorithm (esp how we ignore dblocks with
unmatching lver's.)

As far as I can tell there was never a real problem
with the previous method because of the effect of
lver, and so I don't think this change will effect
the end result in any way.

One minor effect is that the entire lease area
(1MB or 8MB) is now read at the start of a paxos run,
instead of a smaller length that just spanned the
leader block and local dblock.

Signed-off-by: David Teigland <teigland@redhat.com>