sanlock: fix paxos acquire host_id check
The host_id check should be skipped in the case where
the local host_id was the previous paxos lease owner,
but in a previous lockspace generation in which the
paxos lease was not cleanly released.  Checking our
own host_id does nothing and is a waste of time in
this case.

Signed-off-by: David Teigland <teigland@redhat.com>