sanlock: configuring the core dump output
All the setrlimit calls have been moved to a new setup_limits function
that is called regardless the presence of the -U and -G options.
The new RLIMIT_CORE limit is set to allow the core dump to be written
in case of a crash (for more details: man core).

Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>