sanlock: fix aborted command clearing
When clearing the remainder of an aborted
command, non-blocking recv is used to avoid
blocking the main thread, but this means it
needs to do a minimal amount of retrying when
EAGAIN is returned.  It wasn't retrying at all,
so the remainder of the message was handled as
a new command, which caused errors when the
fragment was found to have a bad magic number.

Signed-off-by: David Teigland <teigland@redhat.com>