sanlock/fence_sanlock: request modes FORCE and GRACEFUL
FORCE is new name for KILL_PID.
GRACEFUL is killpath (or SIGTERM of no killpath is set)
SIGUSR1 is removed; a killpath prog should send the signal

fence_sanlock is converted to use killpath to send SIGUSR1.
The killpath program is "fence_sanlockd -1" (-1 is new).

Change sanlock helper to use execvp so we don't need to
know the fence_sanlockd path.

Signed-off-by: David Teigland <teigland@redhat.com>