sanlock: add convert
sanlock_convert is used to convert the mode of an
existing lock, i.e. sh->ex or ex->sh.

Shared is specified by setting the SANLK_RES_SHARED flag as usual
(or :SH suffix on command line.)  Exclusive remains the default,
and is implied by the lack of SANLK_RES_SHARED.

Converting a sh lease to sh, or ex lease to ex will return -EALREADY.

Signed-off-by: David Teigland <teigland@redhat.com>