wdmd: add probe option
The probe option will look for a watchdog device that the
daemon can use.  If a good device is found, the path is printed
and the command exits with 0.  If a good device is not found,
the command exits with 1.

The init script uses this new probe function to determine
if it needs to load the softdog module or not.

Signed-off-by: David Teigland <teigland@redhat.com>