sanlock: log bad delta leases and delta lease owners
Other delta leases in a lockspace are monitored for liveness,
which includes checking and saving the delta lease timestamp.
This adds a check of the delta lease's magic number and
lockspace name, and logs an error if they are not correct.
This will help debugging if a delta lease is corrupted.

This also adds an entry to sanlock.log that shows the owner
information for each delta lease.  This is logged whenever
the sanlock daemon encounters new owner information for a
host in a lockspace.

Signed-off-by: David Teigland <teigland@redhat.com>