sanlock: fix paxos release
Fix a case where a paxos release is clobbered, leaving the lease held on
disk, but not by the host.  Other hosts cannot acquire the lease until the
owner is gone, or acquires and releases it again.

An extra write is added to every paxos release, making a sanlock_release
two writes instead of one.  A host releasing a paxos lease now writes to
its own dblock, zeroing it, then writes the leader record with the owner
field zeroed as before.  (A side effect is that the mbal values in dblocks
restart from zero at each acquire, rather than increasing indefinately.)

The clobbered release only happened when:
- multiple hosts are attempting to acquire the lease at once,
  running the paxos ballot concurrently
- two or more hosts, say A and B, complete the ballot, and both
  write the leader setting the owner to A
- A writes the leader first with itself as owner
- A then quickly releases the lease, writing the leader again,
  with the owner zeroed
- B writes the ballot result to the leader, setting the owner to A
- the leader write from B arrives after the leader write from A
  which released the lease
- so B changes the owner back from zero to A again

This is the correct operation of the algorithm, and there is no clear way
to prevent it (at least without a major rethinking of the implementation).

The solution is to:
1. zero the dblock during release, in addition to the leader owner
2. detect when the leader is clobbered by checking the dblock

A host trying to acquire a lease after the release was clobbered will:
1. see that the lease is owned by A
2. see that the owner of the lease, A, is alive
3. see that last writer of the leader (B) is not the same
   host as the owner of the lease (A)
4. read the dblock of the owner (A) to see if it is zeroed or not
5. if the owner's dblock is zero, it means that the release was
   clobbered as described above
6. the host trying to acquire the lease will go ahead and run
   a new ballot to acquire the lease

This solution uses the write_id part of the leader record which has
previously only been used as a debugging aid.  Using the write_id is not
essential to the solution, but is an optimization.  Without it, the
owner's dblock would need to be read more often (any time acquire sees a
lease held by a live host).

The problem of clobbered releases will occur when a lease is released very
quickly after being acquired.  This quick release is what happens when
acquiring a shared lease, so this problem is most likely to be visible
when multiple hosts all attempt to acquire a shared lease at once.

Signed-off-by: David Teigland <teigland@redhat.com>