sanlock: add_lockspace can connect to wdmd earlier
Currently, add_lockspace will finish acquiring the
delta lease, then connect to wdmd and activate the
wdmd connection for the lockspace.  If wdmd was
not running, the connect fails, and add_lockspace
fails.  If we connect to wdmd before acquiring the
delta lease, we can fail before the lengthy delta
lease delay.