sanlock: default to sanlock user and group if it exists
When the user and group are not specified on the command line
or in the config file, sanlock would previously default to the
"root" user and group.  Now, it will check if the "sanlock"
user and group are available on the system, and if so default
to that instead of root.