sanlock: add acquire flag to avoid waiting
When a paxos lease is requested that was held by
a failed, but not yet expired host, the acquire
routine waits for the lease to expire before
returning.  Add a flag to avoid waiting for the
expiration for programs that want to just retry
instead of block in sanlock_acquire().