sanlock: don't clear the shared flag on disk
When a shared lease is acquired, the shared flag is
written to the mode_block (following the dblock structure),
and then the leader record is released.

In commit e40e1f6e22f9b10f08d53fc7da94f5158d9e4ae8,
releasing the leader record also began to clear the
dblock structure.  Clearing the dblock also wrongly
cleared the mode_block, causing the shared flag to
be lost.  This means that sanlock believes a shared
lease is held, when in fact on disk it is not held,
and can be acquired exclusively by another host.