sanlock: use io_destroy to clear outstanding aio
When a lockspace thread attempts to clear outstanding
aio ops, it was using io_getevents indefinitely to wait
for completions.  Now, it uses io_getevents to clear the
ops for up to 2 minutes, after which it uses io_destroy
to clear them.  io_destroy may have the ability clear
certain ops that will never have a completion.