sanlock: write lease with clear flag
Add the flag SANLK_WRITE_CLEAR to sanlock_write_resource()
which causes the resource lease on disk to be written without
the standard magic number.  This causes a subsequent
sanlock_read_resource() to return the error -223/SANLK_LEADER_MAGIC
indicating that no valid lease exists at that location.

This can be used by an application to indicate that the lease area
on disk is available to be reused for a new resource lease.  (An
application can continue to use a special resource name of its own
choosing to implement the same effect.)