sanlock: improve log messages in paxos code
A lot of messages related to "interesting" conditions
(which may want later inspection) were being written
to /var/log/messages.  Change these to be written only
to sanlock.log or the memory debug buffer.

Log more paxos details so that the behavior of the
algorithm can be more fully reconstructed.