sanlock: fix detection of shared lease
When a host is acquiring a lease and detects that another
host holds it shared, it will check if the host with the
shared lease is dead.  Before the dead-host check, the
shared lease holder may have released its shared lease by
clearing its mode_block.  The host checking the shared lease
needs to check if the shared lease has been released.