sanlock: fix release interference with paxos
The process of releasing a lease on disk involved
first zeroing our own dblock values prior to zeroing
the leader record.  Clearing dblock values can
interfere with an in-progress paxos ballot, and can
result in a second host owning the same lease version
that we released.

The incorrect zeroing of dblocks was added in
commit d6bef45b9716c581d99466280a52a01c9ebe3bf7

The reason for zeroing was to fix an issue with shared
leases here: e40e1f6e22f9b10f08d53fc7da94f5158d9e4ae8
That issue is now resolved by setting a new flag in the
dblock to indicate we are releasing the lease.  The new
flag is set, along with the unchanged dblock values,
before clearing the leader block.