sanlock: fix incorrect error from release
In paxos release, check for different write_id first, rather
than checking for different lver or owner first, because the
different lver or owner is possible with a different write_id.
When there is a write_id other than our own, setting the released
flag in our dblock allows others to acquire the lease, and when
we read the leader record, we can find that it's already held
by another host.  This fix avoids returning an error from paxos
release in this case.