Start rindex tests
Add couple of tests for the new lookup command using the daemon.

To make the tests faster, we use 1 second io timeout. This is much
faster then real usage, but still takes about 3 seconds per test when we
add a lockspace. Need to check how to make this faster.

Unfinished: cover all commands.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>